Change package name; add launcher intent
This makes our app runnable by the front end, at least on our end of
things (it was working for the front end team beforehand, but not
for us).

The launcher intent stuff should be removed for production builds
-- it just makes this test runnable directly through android studio
(instead of only runnable as an APK through the front end).