Get hand/ID/difficulty info from TRIAL intents
This information will be used when sending things to sheets.

This isn't really testable with the current version of the frontend
at present, but once TRIAL intents are ready this should function
correctly (...hopefully -- if not, yell at me or something and I can
fix it).

I also moved the TrialMode code to within our application's source.
Dunno why we have a version of the frontendhelper in the root
of our project -- we should probably delete that eventually?