Fix difficulty display; move some strings to xml
Also changed the help dialog in the case where no hand information
is given (i.e. non-trial modes) -- in this case, the dialog just says
"hand" in lieu of left or right hand. (this is probably only a
temporary fix until we implement our own hand selector for practice
runs, but for the time being it should work well)