Add fully functioning sheets support
This fixes the bugs of 1) the app crashing when writing to the local
sheet and 2) the user being asked for authentication info a ton of
unnecessary times.

Only TODO left here, as far as I can tell, is fixing that bug in which
uploading two images to drive at once just results in two copies of
the same image being uploaded. I suspect there's a race condition
causing this somehow?
And upon further analysis it looks like notifyFinish() is only
triggered after a write() operation and not an upload() operation.
Maybe "concatenate" the path and heatmap bitmaps into two adjacent
images in a single bitmap, and just upload that ...?

In any case: this is pretty great.