Fix bitmap-uploading race condition bug: close #9
Yeah, turns out the reason duplicate images were being
uploaded was a race condition. I got around that by using (abusing?)
notifyFinished() in order to send the data in a three-stage process:
first for the raw trial sheets data, then for the path bitmap, then
for the heatmap bitmap. This process ensures that nothing will
be "trampled on" due to simultaneous upload/write operations from
the same activity.

It might be worth adding a progress wheel/bar/something to indicate
to the user that an intensive operation is in effect during this
process, but I think that's relatively low priority at this time:
personally, I'm just happy we got sheets stuff conclusively working.

Also closing #2 in this commit while I'm at it, since that problem
appears to have been solved by correctly implementing Google Sheets
in the project.